LCA 0.30

+++ LCAimpl.h
@@ -7,7 +7,7 @@ TaxObj* LCA(vector<BlastRes>&, RefTax*, options*);
 //filtering of the blast results
 double filterBlastPrimary(vector<BlastRes>&, options* opt,double&);
 //routine that performs actual LCA matching etc
-TaxObj* LCAcore(const vector<TaxObj*>&, bool &hitRd, double LCAfrac=0.9f, int tdepth=__default_depth);
+TaxObj* LCAcore(const vector<TaxObj*>&, bool &hitRd, double LCAfrac=0.9, int tdepth=__default_depth);
 
 
-vector<TaxObj*> BlastToTax(const vector<BlastRes>& BR, RefTax* RT, options*, float& consPerID);
+vector<TaxObj*> BlastToTax(const vector<BlastRes>& BR, RefTax* RT, options*, double& consPerID);
